# docs/swagger.yaml
definitions:
  models.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.Book:
    properties:
      author:
        type: string
      available_copies:
        type: integer
      id:
        type: integer
      isbn:
        type: string
      title:
        type: string
      total_copies:
        type: integer
    type: object
  models.BorrowRequest:
    properties:
      book_id:
        type: integer
    type: object
  models.Loan:
    properties:
      book_id:
        type: integer
      borrowed_at:
        type: string
      due_date:
        type: string
      id:
        type: integer
      returned_at:
        type: string
      status:
        $ref: '#/definitions/models.LoanStatus'
      user_id:
        type: integer
    type: object
  models.LoanStatus:
    enum:
    - borrowed
    - returned
    type: string
    x-enum-varnames:
    - StatusBorrowed
    - StatusReturned
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      role:
        $ref: '#/definitions/models.UserRole'
    type: object
  models.UserRole:
    enum:
    - admin
    - member
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleMember
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Logs in user with email and password to return a JWT token
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Authenticate user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with member or admin role
      parameters:
      - description: Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Auth
  /books:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Book'
            type: array
      summary: Get all books
      tags:
      - Books
    post:
      consumes:
      - application/json
      description: Adds a new book to the library (Admin only)
      parameters:
      - description: Book Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Book'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Book'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new book
      tags:
      - Books
  /books/{id}:
    get:
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get book by ID
      tags:
      - Books
  /loans/borrow:
    post:
      consumes:
      - application/json
      description: Borrows a book if stock is available and user limit (<3) is not
        exceeded
      parameters:
      - description: Borrow Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BorrowRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Loan'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Borrow a book
      tags:
      - Loans
  /loans/return/{id}:
    post:
      description: Returns an active loan and increments book stock
      parameters:
      - description: Loan ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Loan'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Return a borrowed book
      tags:
      - Loans
swagger: "2.0"